+++ src/components/ui/LinePlotArea/PlotLineOptions.tsx
@@ -9,7 +9,7 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
-import { usePlotStore } from '@/utils/GlobalStates'
+import { usePlotStore } from '@/GlobalStates'
 import { useShallow } from 'zustand/shallow'
 import { Slider } from "@/components/ui/slider"
