+++ src/components/ui/PerformanceMode.tsx
@@ -4,7 +4,7 @@ import { MdOutlineRocketLaunch } from "react-icons/md";
 import { Potato } from './Icons';
 import { FaCarSide } from "react-icons/fa6";
 import { Popover, PopoverTrigger, PopoverContent } from "@/components/ui/popover"
-import { useGlobalStore } from '@/utils/GlobalStates';
+import { useGlobalStore } from '@/GlobalStates';
 import { Button } from "@/components/ui/button"
 
 const icons = {
