+++ src/components/ui/VariablesTable.tsx
@@ -4,7 +4,7 @@ import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import { useGlobalStore } from "@/utils/GlobalStates";
+import { useGlobalStore } from "@/GlobalStates";
 import { useShallow } from "zustand/shallow";
 import MetaDataInfo from "./MainPanel/MetaDataInfo";
 import {
