+++ src/utils/ExportCanvas.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useEffect, useRef } from 'react'
-import { useGlobalStore, useImageExportStore, usePlotStore } from './GlobalStates'
+import { useGlobalStore, useImageExportStore, usePlotStore } from '@/GlobalStates'
 import { useThree } from '@react-three/fiber'
 import { useCSSVariable } from '@/components/ui';
 import * as THREE from 'three'
